# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth": {
      "delete": {
        "operationId": "UserController_delete",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/api/auth/admin/users": {
      "get": {
        "operationId": "AdminController_getUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/admin/block": {
      "patch": {
        "operationId": "AdminController_blockUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/admin/unblock": {
      "patch": {
        "operationId": "AdminController_unblockUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/admin": {
      "patch": {
        "operationId": "AdminController_updateAdmin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/admin/create-user": {
      "post": {
        "operationId": "AdminController_createUserByAdmin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/admin/{id}": {
      "delete": {
        "operationId": "AdminController_deleteUserByAdmin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/admin/add": {
      "patch": {
        "operationId": "AdminController_addAdminRoleToUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/admin/stats": {
      "get": {
        "operationId": "AdminController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/admin/search": {
      "get": {
        "operationId": "AdminController_search",
        "parameters": [
          {
            "name": "query",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/auth/admin/login": {
      "post": {
        "operationId": "AdminController_adminLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/auth/admin/reindex": {
      "put": {
        "operationId": "AdminController_reindexSearchEngine",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/auth/google/success": {
      "get": {
        "operationId": "ProvidersController_googleSuccess",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "surname",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Providers"
        ]
      }
    },
...
    "/api/auth/yandex/success": {
      "get": {
        "operationId": "ProvidersController_yandexSuccess",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "surname",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Providers"
        ]
      }
    },
...
    "/api/marks/admin/{id}": {
      "delete": {
        "operationId": "MarkController_deleteMark",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Marks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/marks/admin/reindex": {
      "put": {
        "operationId": "MarkController_reindexSearhchEngine",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Marks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/categories/admin/reindex": {
      "put": {
        "operationId": "CategoriesController_reindexSearhchEngine",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Incidents API",
    "description": "",
    "version": "0.1.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "in": "header",
        "type": "http"
      }
    },
    "schemas": {
...
      "UserIdDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор пользователя",
            "example": "12345"
          }
        },
        "required": [
          "id"
        ]
      },
      "UpdateAdminDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор администратора",
            "example": "12345"
          },
          "name": {
            "type": "string",
            "description": "Имя администратора",
            "example": "Иван"
          },
          "surname": {
            "type": "string",
            "description": "Фамилия администратора",
            "example": "Иванов"
          },
          "email": {
            "type": "string",
            "description": "Электронная почта администратора",
            "example": "ivan.ivanov@example.com"
          },
          "phone_number": {
            "type": "string",
            "description": "Номер телефона администратора в формате России",
            "example": "+7 900 000-00-00"
          }
        },
        "required": [
          "id",
          "name",
          "surname",
          "email",
          "phone_number"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Электронная почта пользователя",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "Пароль пользователя (минимум 8 символов)",
            "example": "password123"
          },
          "name": {
            "type": "string",
            "description": "Имя пользователя",
            "example": "Иван"
          },
          "surname": {
            "type": "string",
            "description": "Фамилия пользователя",
            "example": "Иванов"
          }
        },
        "required": [
          "email",
          "password",
          "name",
          "surname"
        ]
      },
      "AddAdminDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Идентификатор пользователя, которого нужно назначить администратором",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        },
        "required": [
          "id"
        ]
      },
      "AdminLoginDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Имя администратора для входа в систему",
            "example": "adminUser"
          },
          "password": {
            "type": "string",
            "description": "Пароль администратора",
            "example": "securePassword123"
          },
          "userAgent": {
            "type": "string",
            "description": "Информация о пользовательском агенте, которая может быть предоставлена (не обязательная)",
            "example": "Mozilla/5.0 (Windows NT 10.0; Win64; x64)"
          }
        },
        "required": [
          "name",
          "password"
        ]
      },
...
    }
  }
}